Fixed synchronisation bug
At the synchronisation stage, there was a bug in elements that contain references. Instead of getting the concept of that element/node and checking if that concept contains a reference, I was checking if that node/element contains references. This is bad for cases where a node is empty, as that has no references, so the order of nodes locally was not in sync with that on server. Fixed now.